+++ src/main/java/tekcays_addon/api/metatileentity/predicates/BrickCastingBusPredicates.java
@@ -2,8 +2,8 @@ package tekcays_addon.api.metatileentity.predicates;
 
 import gregtech.api.pattern.TraceabilityPredicate;
 import tekcays_addon.common.blocks.blocks.BlockBrick;
-import tekcays_addon.common.metatileentities.multiblockpart.MetaTileEntityBrickCastingBus;
-import tekcays_addon.common.metatileentities.multiblockpart.MetaTileEntityBrickCastingFluidInput;
+import tekcays_addon.common.metatileentities.multiblockpart.brick.MetaTileEntityBrickCastingBus;
+import tekcays_addon.common.metatileentities.multiblockpart.brick.MetaTileEntityBrickCastingFluidInput;
 
 import static gregtech.api.metatileentity.multiblock.MultiblockControllerBase.metaTileEntities;
 import static tekcays_addon.common.metatileentities.TKCYAMetaTileEntities.*;
