+++ src/main/java/tekcays_addon/common/metatileentities/multi/primitive/MetaTileEntityPrimitiveBath.java
@@ -1,4 +1,4 @@
-package tekcays_addon.common.metatileentities.multi;
+package tekcays_addon.common.metatileentities.multi.primitive;
 
 import codechicken.lib.render.CCRenderState;
 import codechicken.lib.render.pipeline.IVertexOperation;
