+++ src/main/java/tekcays_addon/common/metatileentities/multi/primitive/MetaTileEntityPrimitiveConverter.java
@@ -1,4 +1,4 @@
-package tekcays_addon.common.metatileentities.multi;
+package tekcays_addon.common.metatileentities.multi.primitive;
 
 import gregtech.api.metatileentity.MetaTileEntity;
 import gregtech.api.metatileentity.interfaces.IGregTechTileEntity;
@@ -19,36 +19,36 @@ import net.minecraft.util.text.TextComponentTranslation;
 import net.minecraft.util.text.TextFormatting;
 import net.minecraft.util.text.event.HoverEvent;
 import net.minecraft.world.World;
-import tekcays_addon.common.blocks.TKCYAMetaBlocks;
-import tekcays_addon.common.blocks.blocks.BlockBrick;
 import tekcays_addon.gtapi.capability.containers.IPressureContainer;
 import tekcays_addon.gtapi.metatileentity.multiblock.PressureContainerNoEnergyMultiblockController;
 import tekcays_addon.gtapi.metatileentity.multiblock.TKCYAMultiblockAbility;
 import tekcays_addon.gtapi.recipes.TKCYARecipeMaps;
 import tekcays_addon.gtapi.render.TKCYATextures;
+import tekcays_addon.common.blocks.TKCYAMetaBlocks;
+import tekcays_addon.common.blocks.blocks.BlockBrick;
 
+import java.util.List;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
-import java.util.List;
 
 import static gregtech.api.util.RelativeDirection.*;
 import static tekcays_addon.api.metatileentity.predicates.BrickHatchesPredicates.*;
 
-public class MetaTileEntityPrimitiveRoastOven extends PressureContainerNoEnergyMultiblockController {
+public class MetaTileEntityPrimitiveConverter extends PressureContainerNoEnergyMultiblockController {
 
     private final BlockBrick.BrickType brick;
     private final IBlockState iBlockState;
     private IPressureContainer pressureContainer;
 
-    public MetaTileEntityPrimitiveRoastOven(ResourceLocation metaTileEntityId, BlockBrick.BrickType brickType) {
-        super(metaTileEntityId, TKCYARecipeMaps.ROASTING);
-        this.brick = brickType;
+    public MetaTileEntityPrimitiveConverter(ResourceLocation metaTileEntityId) {
+        super(metaTileEntityId, TKCYARecipeMaps.PRIMITIVE_CONVERTING_RECIPES);
+        this.brick = BlockBrick.BrickType.REINFORCED_BRICK;
         this.iBlockState = TKCYAMetaBlocks.BLOCK_BRICK.getState(brick);
     }
 
     @Override
     public MetaTileEntity createMetaTileEntity(IGregTechTileEntity tileEntity) {
-        return new MetaTileEntityPrimitiveRoastOven(metaTileEntityId, brick);
+        return new MetaTileEntityPrimitiveConverter(metaTileEntityId);
     }
 
     @Override
@@ -103,7 +103,7 @@ public class MetaTileEntityPrimitiveRoastOven extends PressureContainerNoEnergyM
 
     @Override
     public ICubeRenderer getBaseTexture(IMultiblockPart sourcePart) {
-        return TKCYATextures.BRICKS[brick.getTextureId()];
+        return TKCYATextures.BRICKS[BlockBrick.REINFORCED_BRICK];
     }
 
     @Override
@@ -116,7 +116,6 @@ public class MetaTileEntityPrimitiveRoastOven extends PressureContainerNoEnergyM
         return true;
     }
 
-    @Nonnull
     @Override
     protected ICubeRenderer getFrontOverlay() {
         return Textures.COKE_OVEN_OVERLAY;
@@ -126,7 +125,6 @@ public class MetaTileEntityPrimitiveRoastOven extends PressureContainerNoEnergyM
         return iBlockState;
     }
 
-    @Nonnull
     @Override
     protected BlockPattern createStructurePattern() {
         return FactoryBlockPattern.start(RIGHT, FRONT, UP)
@@ -136,12 +134,11 @@ public class MetaTileEntityPrimitiveRoastOven extends PressureContainerNoEnergyM
                 .where('S', selfPredicate())
                 .where('X', states(getCasingState()))
                 .where('Y', states(getCasingState())
-                        .or(getInputBrickFluidHatch(brick).setMinGlobalLimited(1).setMaxGlobalLimited(2))
                         .or(getOutputBrickFluidHatch(brick).setMinGlobalLimited(1).setMaxGlobalLimited(1))
-                        .or(getInputBrickItemBus(brick).setMinGlobalLimited(1).setMaxGlobalLimited(1))
-                        .or(getOutputBrickItemBus(brick).setMinGlobalLimited(1).setMaxGlobalLimited(1))
+                        .or(getInputBrickFluidHatch(brick).setMinGlobalLimited(1).setMaxGlobalLimited(2))
                         .or(abilities(TKCYAMultiblockAbility.PRESSURE_CONTAINER).setMaxGlobalLimited(1, 1)))
-                .where('M', getBrickMuffler(brick))
+                .where('M', metaTileEntities(MetaTileEntities.MUFFLER_HATCH)
+                        .or(getBrickMuffler(brick)))
                 .where('#', air())
                 .build();
     }
