+++ src/main/java/tekcays_addon/common/metatileentities/multi/storage/MetaTileEntityModulableMultiblockTank.java
@@ -1,4 +1,4 @@
-package tekcays_addon.common.metatileentities.multi;
+package tekcays_addon.common.metatileentities.multi.storage;
 
 import codechicken.lib.render.CCRenderState;
 import codechicken.lib.render.pipeline.IVertexOperation;
