+++ src/main/java/tekcays_addon/common/metatileentities/multiblockpart/capabilities/MetaTileEntityVacuumHatch.java
@@ -1,4 +1,4 @@
-package tekcays_addon.common.metatileentities.multiblockpart;
+package tekcays_addon.common.metatileentities.multiblockpart.capabilities;
 
 import codechicken.lib.render.CCRenderState;
 import codechicken.lib.render.pipeline.IVertexOperation;
